Moved iterators into `src/utils/iterators`.
