+++ src/utils/iterators/collatz.rs
@@ -1,0 +1,28 @@
+/// Collatz sequence iterator.
+pub struct Collatz {
+    num: i64,
+    done: bool,
+}
+impl Collatz {
+    pub fn new(num: i64) -> Collatz {
+        Collatz { num, done: false }
+    }
+}
+impl Iterator for Collatz {
+    type Item = i64;
+    fn next(&mut self) -> Option<i64> {
+        if self.done {
+            return None;
+        }
+        let num = self.num;
+        self.num = if self.num % 2 == 0 {
+            self.num / 2
+        } else {
+            3 * self.num + 1
+        };
+        self.done = num == 1;
+        Some(num)
+    }
+}
+
+
