+++ src/utils/iterators/digits.rs
@@ -1,0 +1,24 @@
+/// Digits iterator. Generates the decimal digits of a number from least
+/// significant to most significant. Positive numbers only!
+pub struct Digits {
+    num: i64,
+}
+impl Digits {
+    pub fn new(num: i64) -> Digits {
+        Digits { num }
+    }
+}
+impl Iterator for Digits {
+    type Item = i64;
+    fn next(&mut self) -> Option<i64> {
+        if self.num == 0 {
+            None
+        } else {
+            let digit = self.num % 10;
+            self.num /= 10;
+            Some(digit)
+        }
+    }
+}
+
+
