+++ src/utils/iterators/divisors.rs
@@ -1,0 +1,45 @@
+use crate::utils;
+
+/// Divisors iterator. Generates all divisors of a number in an unspecified
+/// order. Positive numbers only!
+pub struct Divisors {
+    dividend: i64,
+    limit: i64,
+    current: i64,
+    other: i64,
+}
+impl Divisors {
+    pub fn new(dividend: i64) -> Divisors {
+        Divisors {
+            dividend,
+            limit: utils::isqrt(dividend),
+            current: 0,
+            other: 0,
+        }
+    }
+}
+impl Iterator for Divisors {
+    type Item = i64;
+    fn next(&mut self) -> Option<i64> {
+        if self.other > 0 {
+            let other = self.other;
+            self.other = 0;
+            return Some(other);
+        }
+        loop {
+            self.current += 1;
+            if self.current > self.limit {
+                return None;
+            }
+            if self.dividend % self.current == 0 {
+                let other = self.dividend / self.current;
+                if other != self.current {
+                    self.other = other;
+                }
+                return Some(self.current);
+            }
+        }
+    }
+}
+
+
