+++ src/utils/iterators/polygonal.rs
@@ -1,0 +1,49 @@
+
+/// Triangular, pentagonal, hexagonal, etc. number iterator. Specify the
+/// number of sides of the polygon as the argument to the constructor.
+pub struct Polygonal {
+    increment: i64,
+    offset: i64,
+    num: i64,
+}
+impl Polygonal {
+    pub fn new(sides: i64) -> Polygonal {
+        Polygonal {
+            increment: sides - 2,
+            offset: 1,
+            num: 0,
+        }
+    }
+    /// Find the index at which the given number would appear in a sequence of
+    /// polygonal numbers (if it is a polygonal number).
+    ///
+    /// * `sides` - Number of sides of the polygon the sequence is based on.
+    /// * `num` - Number whose index is to be found.
+    ///
+    /// -> Index.
+    pub fn invert(sides: i64, num: i64) -> Result<i64, i64> {
+        // A polygonal number is a quadratic function of the index it appears
+        // at. Solve for the positive root of the corresponding quadratic
+        // equation.
+        let a = sides - 2;
+        let b = 4 - sides;
+        let c = -2 * num;
+        let discriminant = b.pow(2) - 4 * a * c;
+        let idx = (-b as f64 + (discriminant as f64).sqrt()) / (2.0 * a as f64);
+        let idx_rounded = idx.round() as i64;
+        if idx.fract() == 0.0 {
+            Ok(idx_rounded)
+        } else {
+            Err(idx_rounded)
+        }
+    }
+}
+impl Iterator for Polygonal {
+    type Item = i64;
+    fn next(&mut self) -> Option<i64> {
+        self.num += self.offset;
+        self.offset += self.increment;
+        Some(self.num)
+    }
+}
+
